requirements.txt
pymupdf==1.26.7
pillow==12.0.0
jsonschema==4.25.1
